Fixed bug with varargs

+++ arch/ARM/ARMInstPrinter.c
@@ -625,7 +625,7 @@ void printAM3PreOrOffsetIndexOp(MCInst *MI, unsigned Op, SStream *O,
 		add_cs_detail(MI, \
 			      CONCAT(ARM_OP_GROUP_AddrMode3Operand, \
 				     AlwaysPrintImm0), \
-			      Op, AlwaysPrintImm0); \
+			      Op, (uint64_t)AlwaysPrintImm0); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (Op)); \
 		if (!MCOperand_isReg(MO1)) { \
 			printOperand(MI, Op, O); \
@@ -699,7 +699,7 @@ void printPostIdxImm8s4Operand(MCInst *MI, unsigned OpNum, SStream *O)
 	{ \
 		add_cs_detail( \
 			MI, CONCAT(ARM_OP_GROUP_MveAddrModeRQOperand, shift), \
-			OpNum, shift); \
+			OpNum, (uint64_t)shift); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (OpNum)); \
 		MCOperand *MO2 = MCInst_getOperand(MI, (OpNum + 1)); \
 \
@@ -735,7 +735,7 @@ DEFINE_printMveAddrModeRQOperand(0) DEFINE_printMveAddrModeRQOperand(3)
 		add_cs_detail(MI, \
 			      CONCAT(ARM_OP_GROUP_AddrMode5Operand, \
 				     AlwaysPrintImm0), \
-			      OpNum, AlwaysPrintImm0); \
+			      OpNum, (uint64_t)AlwaysPrintImm0); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (OpNum)); \
 		MCOperand *MO2 = MCInst_getOperand(MI, (OpNum + 1)); \
 \
@@ -763,7 +763,7 @@ DEFINE_printAddrMode5Operand(false) DEFINE_printAddrMode5Operand(true)
 		add_cs_detail(MI, \
 			      CONCAT(ARM_OP_GROUP_AddrMode5FP16Operand, \
 				     AlwaysPrintImm0), \
-			      OpNum, AlwaysPrintImm0); \
+			      OpNum, (uint64_t)AlwaysPrintImm0); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (OpNum)); \
 		MCOperand *MO2 = MCInst_getOperand(MI, (OpNum + 1)); \
 \
@@ -1166,7 +1166,7 @@ void printPCLabel(MCInst *MI, unsigned OpNum, SStream *O)
 						 SStream *O) \
 	{ \
 		add_cs_detail(MI, CONCAT(ARM_OP_GROUP_AdrLabelOperand, scale), \
-			      OpNum, scale); \
+			      OpNum, (uint64_t)scale); \
 		MCOperand *MO = MCInst_getOperand(MI, (OpNum)); \
 \
 		if (MCOperand_isExpr(MO)) { \
@@ -1320,7 +1320,7 @@ void printT2SOOperand(MCInst *MI, unsigned OpNum, SStream *O)
 		add_cs_detail(MI, \
 			      CONCAT(ARM_OP_GROUP_AddrModeImm12Operand, \
 				     AlwaysPrintImm0), \
-			      OpNum, AlwaysPrintImm0); \
+			      OpNum, (uint64_t)AlwaysPrintImm0); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (OpNum)); \
 		MCOperand *MO2 = MCInst_getOperand(MI, (OpNum + 1)); \
 \
@@ -1359,7 +1359,7 @@ DEFINE_printAddrModeImm12Operand(false) DEFINE_printAddrModeImm12Operand(true)
 		add_cs_detail(MI, \
 			      CONCAT(ARM_OP_GROUP_T2AddrModeImm8Operand, \
 				     AlwaysPrintImm0), \
-			      OpNum, AlwaysPrintImm0); \
+			      OpNum, (uint64_t)AlwaysPrintImm0); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (OpNum)); \
 		MCOperand *MO2 = MCInst_getOperand(MI, (OpNum + 1)); \
 \
@@ -1394,7 +1394,7 @@ DEFINE_printAddrModeImm12Operand(false) DEFINE_printAddrModeImm12Operand(true)
 		add_cs_detail(MI, \
 			      CONCAT(ARM_OP_GROUP_T2AddrModeImm8s4Operand, \
 				     AlwaysPrintImm0), \
-			      OpNum, AlwaysPrintImm0); \
+			      OpNum, (uint64_t)AlwaysPrintImm0); \
 		MCOperand *MO1 = MCInst_getOperand(MI, (OpNum)); \
 		MCOperand *MO2 = MCInst_getOperand(MI, (OpNum + 1)); \
 \
@@ -1818,7 +1818,7 @@ void printVectorListFourSpaced(MCInst *MI, unsigned OpNum, SStream *O)
 						 SStream *O) \
 	{ \
 		add_cs_detail(MI, CONCAT(ARM_OP_GROUP_MVEVectorList, NumRegs), \
-			      OpNum, NumRegs); \
+			      OpNum, (uint64_t)NumRegs); \
 		unsigned Reg = \
 			MCOperand_getReg(MCInst_getOperand(MI, (OpNum))); \
 		const char *Prefix = "{"; \
@@ -1841,7 +1841,7 @@ DEFINE_printMVEVectorList(2) DEFINE_printMVEVectorList(4)
 			MI, \
 			CONCAT(CONCAT(ARM_OP_GROUP_ComplexRotationOp, Angle), \
 			       Remainder), \
-			OpNo, Angle, Remainder); \
+			OpNo, (uint64_t)Angle, (uint64_t)Remainder); \
 		unsigned Val = \
 			MCOperand_getImm(MCInst_getOperand(MI, (OpNo))); \
 		SStream_concat(O, "#%d", (Val * Angle) + Remainder); \
